// src/modules/activity/type/activity.types.ts
export enum ActivityAction {
  TASK_CREATED = 'TASK_CREATED',
  TASK_UPDATED = 'TASK_UPDATED',
  STATUS_CHANGED = 'STATUS_CHANGED',
  COMMENT_ADDED = 'COMMENT_ADDED',
  ASSIGNEE_CHANGED = 'ASSIGNEE_CHANGED',
  REMOVED_FROM_SPRINT = 'REMOVED_FROM_SPRINT',

  PROJECT_CREATED = 'PROJECT_CREATED',
  PROJECT_DELETED = 'PROJECT_DELETED',
  COLUMN_DELETED = 'COLUMN_DELETED',
  PROJECT_UPDATED = 'PROJECT UPDATED',
  ROLE_CHANGED = 'ROLE_CHANGED',
  MEMBER_ADDED = 'MEMBER_ADDED',
  MEMBER_REMOVED = 'MEMBER_REMOVED',
}